requirements.txt
Flask
openai
elevenlabs-sdk